Move Config to SPI
