+++ extensions/http/jetty/src/main/java/org/eclipse/dataspaceconnector/extension/jetty/JettyConfiguration.java
@@ -15,7 +15,7 @@
 package org.eclipse.dataspaceconnector.extension.jetty;
 
 import org.eclipse.dataspaceconnector.spi.EdcSetting;
-import org.eclipse.dataspaceconnector.spi.system.Config;
+import org.eclipse.dataspaceconnector.spi.system.configuration.Config;
 
 import java.util.AbstractMap;
 import java.util.HashMap;
