+++ extensions/http/jetty/src/test/java/org/eclipse/dataspaceconnector/extension/jetty/JettyConfigurationTest.java
@@ -15,7 +15,7 @@
 package org.eclipse.dataspaceconnector.extension.jetty;
 
 import org.assertj.core.api.ThrowableAssert;
-import org.eclipse.dataspaceconnector.core.config.ConfigFactory;
+import org.eclipse.dataspaceconnector.spi.system.configuration.ConfigFactory;
 import org.junit.jupiter.api.Test;
 
 import java.util.Map;
